make health card not a card
