+++ client/src/app/game/actions/mobile-actions.tsx
@@ -110,7 +110,6 @@ const MobileActions = ({onClick, localGameState, mobile, id}: Props) => {
 					type={'single_use'}
 					onClick={handleClick}
 					playerId={playerId}
-					statusEffects={localGameState.statusEffects}
 				/>
 			</div>
 		)
