+++ client/src/app/game/board/board-slot.tsx
@@ -7,10 +7,6 @@ import HermitCard from 'common/cards/base/hermit-card'
 import EffectCard from 'common/cards/base/effect-card'
 import SingleUseCard from 'common/cards/base/single-use-card'
 import ItemCard from 'common/cards/base/item-card'
-import HealthCard from 'common/cards/base/health-card'
-import {StatusEffectT} from 'common/types/game-state'
-import StatusEffect from 'components/status-effects/status-effect'
-import {STATUS_EFFECT_CLASSES} from 'common/status-effects'
 import {SlotTypeT} from 'common/types/cards'
 import {useSelector} from 'react-redux'
 import {
@@ -19,8 +15,6 @@ import {
 	getPickRequestPickableSlots,
 	getSelectedCard,
 } from 'logic/game/game-selectors'
-import {getLocalPlayerState} from 'server/src/utils/state-gen'
-import {slot} from 'common/slot'
 
 export type SlotProps = {
 	type: SlotTypeT
@@ -32,78 +26,17 @@ export type SlotProps = {
 	rowState?: RowState
 	active?: boolean
 	cssId?: string
-	statusEffects: Array<StatusEffectT>
 }
-const Slot = ({
-	type,
-	rowIndex,
-	index,
-	playerId,
-	onClick,
-	card,
-	rowState,
-	active,
-	cssId,
-	statusEffects,
-}: SlotProps) => {
+const Slot = ({type, rowIndex, index, playerId, onClick, card, active, cssId}: SlotProps) => {
 	const cardsCanBePlacedIn = useSelector(getCardsCanBePlacedIn)
 	const pickRequestPickableCard = useSelector(getPickRequestPickableSlots)
 	const selectedCard = useSelector(getSelectedCard)
 	const localGameState = useSelector(getGameState)
 
 	let cardInfo = card?.cardId
-		? (CARDS[card.cardId] as HermitCard | EffectCard | SingleUseCard | ItemCard | HealthCard)
+		? (CARDS[card.cardId] as HermitCard | EffectCard | SingleUseCard | ItemCard)
 		: null
-	if (type === 'health' && rowState?.health) {
-		cardInfo = new HealthCard({
-			id: 'health',
-			name: 'Health Card',
-			rarity: 'common',
-			health: rowState.health,
-		})
-	}
 
-	const renderStatusEffects = (cleanedStatusEffects: StatusEffectT[]) => {
-		return (
-			<div className={css.statusEffectContainer}>
-				{cleanedStatusEffects.map((a) => {
-					const statusEffect = STATUS_EFFECT_CLASSES[a.statusEffectId]
-					if (!statusEffect || !statusEffect.visible) return null
-					if (statusEffect.damageEffect == true) return null
-					return <StatusEffect statusEffect={statusEffect} duration={a.duration} />
-				})}
-			</div>
-		)
-	}
-	const renderDamageStatusEffects = (cleanedStatusEffects: StatusEffectT[] | null) => {
-		return (
-			<div className={css.damageStatusEffectContainer}>
-				{cleanedStatusEffects
-					? cleanedStatusEffects.map((a) => {
-							const statusEffect = STATUS_EFFECT_CLASSES[a.statusEffectId]
-							if (!statusEffect || !statusEffect.visible) return null
-							if (statusEffect.damageEffect == false) return null
-							return <StatusEffect statusEffect={statusEffect} />
-					  })
-					: null}
-			</div>
-		)
-	}
-
-	const hermitStatusEffects = Array.from(
-		new Set(
-			statusEffects
-				.filter((a) => rowState?.hermitCard && a.targetInstance == rowState.hermitCard.cardInstance)
-				.map((a) => a) || []
-		)
-	)
-	const effectStatusEffects = Array.from(
-		new Set(
-			statusEffects.filter(
-				(a) => rowState?.effectCard && a.targetInstance == rowState.effectCard.cardInstance
-			) || []
-		)
-	)
 	const frameImg = type === 'hermit' ? '/images/game/frame_glow.png' : '/images/game/frame.png'
 
 	const getPickableSlots = () => {
@@ -166,18 +99,8 @@ const Slot = ({
 			{cardInfo ? (
 				<div className={css.cardWrapper}>
 					<Card card={cardInfo} />
-					{type === 'health'
-						? renderStatusEffects(hermitStatusEffects)
-						: type === 'effect'
-						? renderStatusEffects(effectStatusEffects)
-						: null}
-					{type === 'health'
-						? renderDamageStatusEffects(hermitStatusEffects)
-						: type === 'effect'
-						? renderDamageStatusEffects(effectStatusEffects)
-						: renderDamageStatusEffects(null)}
 				</div>
-			) : type === 'health' ? null : (
+			) : (
 				<img draggable="false" className={css.frame} src={frameImg} />
 			)}
 		</div>
