+++ client/src/components/card/health-card-svg.module.scss
@@ -29,8 +29,10 @@
 
 .type {
 	font-size: 72px;
+	font-family: BangersBold, Helvetica, serif;
 }
 
 .health {
 	font-size: 148px;
+	font-family: BangersBold, Helvetica, serif;
 }
