+++ client/src/components/card/health-card-svg.tsx
@@ -1,11 +1,6 @@
-import HealthCard from 'common/cards/base/health-card'
 import {useState, useRef, useMemo, useEffect, memo} from 'react'
 import css from './health-card-svg.module.scss'
 import classnames from 'classnames'
-
-export type HealthCardProps = {
-	card: HealthCard
-}
 
 function useCountdownAnimation(value: number, duration = 500) {
 	const [displayValue, setDisplayValue] = useState(value)
@@ -43,8 +38,8 @@ function useCountdownAnimation(value: number, duration = 500) {
 	return Math.round(displayValue)
 }
 
-const HealthCardModule = memo(({card}: HealthCardProps) => {
-	const displayHealth = useCountdownAnimation(card.health)
+const HealthDisplayModule = memo(({health}: {health: number}) => {
+	const displayHealth = useCountdownAnimation(health)
 
 	return (
 		<svg
@@ -96,4 +91,4 @@ const HealthCardModule = memo(({card}: HealthCardProps) => {
 	)
 })
 
-export default HealthCardModule
+export default HealthDisplayModule
